Split GAMS solver into separate module.
